PM / devfreq: Add a governor white list
Many of the devfreq governors compiled in are unstable/incomplete, so we
don't want users switching to them. Compiling out the unused governors is
messy because many other drivers link to them, and there are not proper

Add a white list of the only governors intended for use to the governor
sysfs node in order to prevent users from enabling governors that unstable.

Change-Id: I2829b0dcbe0e84236f8ead039b0e028b73834e5f
Signed-off-by: Sultanxda <sultanxda@gmail.com>